HNSW graphs serialization

+++ TestHNSW/HNSWDemo/Tests/LALTest.cs
@@ -12,6 +12,8 @@ namespace HNSWDemo.Tests
     {
         private const int count = 20000;
         private const int dimensionality = 128;
+        private const string _graphFileCachee = @"lal_test_graph.bin";
+        private const string _mapFileCachee = @"lal_test_map.bin";
 
         public void Run()
         {
@@ -27,52 +29,66 @@ namespace HNSWDemo.Tests
                 samples[c].Add(p);
             }
 
+            SplittedLALGraph worlds;
+            HNSWMappers<long> mappers;
 
-            var worlds = new SplittedLALGraph();
-            var mappers = new HNSWMappers<long>(l => (int)Math.Abs(l.GetHashCode() % moda));
-
-            var worlds_dict = new Dictionary<int, SmallWorld<float[]>>();
-            var maps_dict = new Dictionary<int, HNSWMap<long>>();
-
-            foreach (var p in samples)
+            if (File.Exists(_graphFileCachee) && File.Exists(_mapFileCachee))
             {
-                var c = p.Key;
-                if (worlds_dict.ContainsKey(c) == false)
-                {
-                    worlds_dict.Add(c, new SmallWorld<float[]>(options));
-                }
-                if (maps_dict.ContainsKey(c) == false)
-                {
-                    maps_dict.Add(c, new HNSWMap<long>());
-                }
-                var w = worlds_dict[c];
-                var m = maps_dict[c];
-                var ids = w.AddItems(p.Value.Select(i => i.Item1));
-
-                for (int i = 0; i < ids.Length; i++)
-                {
-                    m.Append(p.Value[i].Item2.Number, ids[i]);
-                }
+                worlds = new SplittedLALGraph(_graphFileCachee);
+                mappers = new HNSWMappers<long>(_mapFileCachee, l => (int)Math.Abs(l.GetHashCode() % moda));
             }
-
-            var name = Guid.NewGuid().ToString();
-            foreach (var p in samples)
+            else
             {
-                var c = p.Key;
-                var w = worlds_dict[c];
-                var m = maps_dict[c];
 
-                using (var s = File.Create(name))
+                worlds = new SplittedLALGraph();
+                mappers = new HNSWMappers<long>(l => (int)Math.Abs(l.GetHashCode() % moda));
+
+                var worlds_dict = new Dictionary<int, SmallWorld<float[]>>();
+                var maps_dict = new Dictionary<int, HNSWMap<long>>();
+
+                foreach (var p in samples)
                 {
-                    w.Serialize(s);
+                    var c = p.Key;
+                    if (worlds_dict.ContainsKey(c) == false)
+                    {
+                        worlds_dict.Add(c, new SmallWorld<float[]>(options));
+                    }
+                    if (maps_dict.ContainsKey(c) == false)
+                    {
+                        maps_dict.Add(c, new HNSWMap<long>());
+                    }
+                    var w = worlds_dict[c];
+                    var m = maps_dict[c];
+                    var ids = w.AddItems(p.Value.Select(i => i.Item1));
+
+                    for (int i = 0; i < ids.Length; i++)
+                    {
+                        m.Append(p.Value[i].Item2.Number, ids[i]);
+                    }
                 }
-                using (var s = File.OpenRead(name))
+
+                var name = Guid.NewGuid().ToString();
+                foreach (var p in samples)
                 {
-                    var l = LALGraph.FromHNSWGraph<float[]>(s);
-                    worlds.Append(l, c);
+                    var c = p.Key;
+                    var w = worlds_dict[c];
+                    var m = maps_dict[c];
+
+                    using (var s = File.Create(name))
+                    {
+                        w.Serialize(s);
+                    }
+                    using (var s = File.OpenRead(name))
+                    {
+                        var l = LALGraph.FromHNSWGraph<float[]>(s);
+                        worlds.Append(l, c);
+                    }
+                    File.Delete(name);
+                    mappers.Append(m, c);
                 }
-                File.Delete(name);
-                mappers.Append(m, c);
+
+                worlds.Save(_graphFileCachee);
+                mappers.Save(_mapFileCachee);
             }
 
             var entries = new long[10];
@@ -80,7 +96,6 @@ namespace HNSWDemo.Tests
             {
                 entries[i] = persons[DefaultRandomGenerator.Instance.Next(0, persons.Count - 1)].Item2.Number;
             }
-
             var contexts = mappers.CreateContext(null, entries);
             var result = worlds.KNearest(5000, contexts);
